Make configuration an instance
So different clients can be initialized with different configuration
